+++ src/settings.ts
@@ -1,9 +1,19 @@
 /**
  * This is the name of the platform that users will use to register the plugin in the Homebridge config.json
  */
-export const PLATFORM_NAME = 'HoneywellThermostatPlatform';
+export const PLATFORM_NAME = 'HoneywellHome';
 
 /**
  * This must match the name of your plugin as defined the package.json
  */
 export const PLUGIN_NAME = 'homebridge-honeywell-home';
+
+/**
+ * This is the main url used to access honeywell API
+ */
+export const HoneywellUrl = 'https://api.honeywell.com/';
+
+/**
+ * This is the url used to access UI Login to honeywell API
+ */
+export const UIurl = 'https://homebridge-honeywell.iot.oz.nu/user/refresh';